Remove unnecessary hashing from list of hashes
I don't see any reason this needs to be here, these are just acting as
keys for a dictionary, and a byte string of hashes will work just as
well as a key, without re-hashing it into a different string.  A
grouped tuple of such byte strings is also a unique identifier, which
Python dict implicitly creates a hash for lookup anyway.